refactor(file): Model폴더에 있는 2개의 파일 Controlloer파일을 만들어 이동
- 이동 파일: isCommaSemicolonConditionCheck.js, isCustomConditionChekc.js
